Update MetaphorandVerbnet.py
Obtained a method to find the verb sense of the verb used in run. If metaphor then it spits out all senses of run. However, if literal, gives the usage of run in the script. However, I can't get it to actually print metaphor or literal, shown on the last line.